SierraItemUpdater with deepstore items
